+++ spec/recurly/plan_spec.rb
@@ -8,7 +8,7 @@ describe Plan do
       :unit_amount_in_cents      => 79_00,
       :description               => "The Gold Plan is for folks who love gold.",
       :accounting_code           => "gold_plan_acc_code",
-      :setup_fee_accounting_code => "Setup Fee AC",
+      :setup_fee_accounting_code => "setup_fee_ac",
       :setup_fee_in_cents        => 60_00,
       :plan_interval_length      => 1,
       :plan_interval_unit        => 'months',
@@ -25,7 +25,7 @@ describe Plan do
 <plan_code>gold</plan_code>\
 <plan_interval_length>1</plan_interval_length>\
 <plan_interval_unit>months</plan_interval_unit>\
-<setup_fee_accounting_code>Setup Fee AC</setup_fee_accounting_code>\
+<setup_fee_accounting_code>setup_fee_ac</setup_fee_accounting_code>\
 <setup_fee_in_cents>\
 <USD>6000</USD>\
 </setup_fee_in_cents>\